Fix link to service desk
The old link never worked for me. The login page it would take me to never seemed to work. It would ack that I exist but said I didn't have an account on the cncf site. I wonder if that login is more for admins. The new link does seem to work. Notice that it's a totally different page - which, again, makes me think the wrong access login portal was being used.